refactor(deliner): use class for consistency
