+++ src/constants/deliner.ts
@@ -1,0 +1,3 @@
+export const commandIds = {
+  deleteAll: 'veco.deliner.deleteAll',
+} as const
